+++ src/components/ui/typings/mp4.d.ts
@@ -1,0 +1,4 @@
+declare module "*.mp4" {
+  const value: any;
+  export = value;
+}