+++ src/lib/helpers/trimlines.js
@@ -1,14 +1,30 @@
-const LINE_MAX = 3;
+const LINE_MAX = 4;
 const LINE_MIN = 2;
 const CHARS_PER_LINE = 18;
 const CHARS_MAX_TOTAL = 18 + 18 + 16;
 
-const varRegex = new RegExp("\\$[VLT]?[0-9]+\\$", "g");
-const varCharRegex = new RegExp("#[VLT]?[0-9]+#", "g");
-const commandRegex = new RegExp("\\!S[0-5]\\!", "g");
+export const varRegex = new RegExp("\\$[VLT]?[0-9]+\\$", "g");
+export const varCharRegex = new RegExp("#[VLT]?[0-9]+#", "g");
+export const commandRegex = new RegExp("\\!S[0-5]\\!", "g");
+export const newlineRegex = new RegExp("\n", "g");
+
+export const dummyText = (text) => {
+  return text
+    .replace(varRegex, "$$$$$")
+    .replace(varCharRegex, "#")
+    .replace(commandRegex, "");
+};
 
 export const lineLength = (line) => {
   return line
+    .replace(varRegex, "255")
+    .replace(varCharRegex, "C")
+    .replace(commandRegex, "").length;
+};
+
+export const totalLength = (line) => {
+  return line
+    .replace(newlineRegex, "")
     .replace(varRegex, "255")
     .replace(varCharRegex, "C")
     .replace(commandRegex, "").length;
@@ -27,7 +43,8 @@ const cropLine = (line, maxLength) => {
 const trimlines = (
   string,
   maxCharsPerLine = CHARS_PER_LINE,
-  maxLines = LINE_MAX
+  maxLines = LINE_MAX,
+  maxTotal = CHARS_MAX_TOTAL
 ) => {
   let lengthCount = 0;
 
@@ -67,8 +84,8 @@ const trimlines = (
       .split("\n")
       .map((line) => {
         lengthCount += lineLength(line);
-        if (lengthCount > CHARS_MAX_TOTAL) {
-          const amountOver = lengthCount - CHARS_MAX_TOTAL;
+        if (lengthCount > maxTotal) {
+          const amountOver = lengthCount - maxTotal;
           return line.substring(0, line.length - amountOver);
         }
         return line;